+++ jupyter_server_proxy/standalone/__init__.py
@@ -1,0 +1,9 @@
+from .app import StandaloneProxyServer
+
+
+def main():
+    StandaloneProxyServer.launch_instance()
+
+
+if __name__ == "__main__":
+    main()